+++ src/FinViet.Infrastructure/Services/IncomeAllocationService.cs
@@ -76,9 +76,9 @@ public class IncomeAllocationService : IIncomeAllocationService
     public async Task<IncomeAllocationEntryDto> ScheduleNextMonthAsync(
         Guid customerId,
         decimal monthlyIncome,
-        int needsPct,
-        int wantsPct,
-        int savingsPct,
+        decimal needsPct,
+        decimal wantsPct,
+        decimal savingsPct,
         CancellationToken cancellationToken = default)
     {
         var exists = await _db.Customers.AnyAsync(c => c.CustomerId == customerId, cancellationToken);
